Refactor dashboard view into dedicated components
